bugfix/AB#82677_ABC-popup-is-not-opening-when-clicking-available-grid-fields-on-question-of-type-resources refactor: move config display fields modal component to the survey components folder as it's only used there to keep all related components ordered and at hand for the same group fix: use ngZone to open the dialog within the survey from a custom angular question component in order to keep the change detection sync with the related angular question inside the surveyjs fix: add some ui improvements for the available grid fields button to display an spinner while loading resource refactor: remove not needed css file
