+++ libs/shared/src/lib/survey/components/resource-available-fields/config-display-grid-fields-modal/config-display-grid-fields-modal.component.ts
@@ -1,5 +1,4 @@
 import {
-  ChangeDetectorRef,
   Component,
   Inject,
   OnInit,
@@ -7,18 +6,18 @@ import {
   ViewContainerRef,
 } from '@angular/core';
 import { DIALOG_DATA } from '@angular/cdk/dialog';
-import { QueryBuilderService } from '../../services/query-builder/query-builder.service';
 import { UntypedFormGroup } from '@angular/forms';
-import { createQueryForm } from '../query-builder/query-builder-forms';
 import { CommonModule } from '@angular/common';
 import { SpinnerModule } from '@oort-front/ui';
-import { QueryBuilderModule } from '../query-builder/query-builder.module';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { TranslateModule } from '@ngx-translate/core';
 import { DialogModule } from '@oort-front/ui';
 import { ButtonModule } from '@oort-front/ui';
 import { takeUntil } from 'rxjs';
-import { UnsubscribeComponent } from '../utils/unsubscribe/unsubscribe.component';
+import { QueryBuilderModule } from '../../../../components/query-builder/query-builder.module';
+import { QueryBuilderService } from '../../../../services/query-builder/query-builder.service';
+import { createQueryForm } from '../../../../components/query-builder/query-builder-forms';
+import { UnsubscribeComponent } from '../../../../components/utils/unsubscribe/unsubscribe.component';
 
 /**
  * Interface that describes the structure of the data shown in the dialog
@@ -29,7 +28,7 @@ interface DialogData {
 }
 
 /**
- * This component is used in the grids to display a modal to configurate the fields in the grid
+ * This component is used in the grids to display a modal to configure the fields in the grid
  */
 @Component({
   standalone: true,
@@ -45,7 +44,6 @@ interface DialogData {
   ],
   selector: 'shared-config-display-grid-fields-modal',
   templateUrl: './config-display-grid-fields-modal.component.html',
-  styleUrls: ['./config-display-grid-fields-modal.component.css'],
 })
 export class ConfigDisplayGridFieldsModalComponent
   extends UnsubscribeComponent
@@ -69,8 +67,7 @@ export class ConfigDisplayGridFieldsModalComponent
    */
   constructor(
     @Inject(DIALOG_DATA) public data: DialogData,
-    private queryBuilder: QueryBuilderService,
-    private cdr: ChangeDetectorRef
+    private queryBuilder: QueryBuilderService
   ) {
     super();
   }
@@ -79,26 +76,23 @@ export class ConfigDisplayGridFieldsModalComponent
     this.queryBuilder.availableQueries$
       .pipe(takeUntil(this.destroy$))
       .subscribe((res) => {
-        this.setUpDialogData(res);
-      });
-  }
+        if (res.length > 0) {
+          const hasDataForm = this.data.form !== null;
+          const queryName = hasDataForm
+            ? this.data.form.value.name
+            : this.queryBuilder.getQueryNameFromResourceName(
+                this.data.resourceName
+              );
 
-  private setUpDialogData(response: any) {
-    if (response.length > 0) {
-      const hasDataForm = this.data.form !== null;
-      const queryName = hasDataForm
-        ? this.data.form.value.name
-        : this.queryBuilder.getQueryNameFromResourceName(
-            this.data.resourceName
-          );
-      this.form = createQueryForm({
-        name: queryName,
-        fields: hasDataForm ? this.data.form.value.fields : [],
-        sort: hasDataForm ? this.data.form.value.sort : {},
-        filter: hasDataForm ? this.data.form.value.filter : {},
+          this.form = createQueryForm({
+            name: queryName,
+            fields: hasDataForm ? this.data.form.value.fields : [],
+            sort: hasDataForm ? this.data.form.value.sort : {},
+            filter: hasDataForm ? this.data.form.value.filter : {},
+          });
+
+          this.loading = false;
+        }
       });
-      this.loading = false;
-      this.cdr.detectChanges();
-    }
   }
 }
