+++ libs/shared/src/lib/survey/components/resource-available-fields/resource-available-fields.component.ts
@@ -1,6 +1,7 @@
 import {
   ChangeDetectorRef,
   Component,
+  NgZone,
   OnDestroy,
   ViewContainerRef,
 } from '@angular/core';
@@ -57,13 +58,15 @@ export class ResourceAvailableFieldsComponent
    * @param {UntypedFormBuilder} fb - Angular - Form builder utilities
    * @param {Dialog} dialog - Angular CDK - This is the Dialog service that is used to handle cdk dialogs
    * @param {Apollo} apollo - Apollo - This is the Apollo service that we'll use to make our GraphQL queries.
+   * @param {NgZone} ngZone - Angular - This is angular injectable service for executing work inside or outside of the Angular zone.
    */
   constructor(
     changeDetectorRef: ChangeDetectorRef,
     viewContainerRef: ViewContainerRef,
     private fb: UntypedFormBuilder,
     private dialog: Dialog,
-    private apollo: Apollo
+    private apollo: Apollo,
+    private ngZone: NgZone
   ) {
     super(changeDetectorRef, viewContainerRef);
   }
@@ -71,12 +74,14 @@ export class ResourceAvailableFieldsComponent
   /**
    * Open a modal for the selected resource with all the available fields
    */
-  openResourceFieldsModal() {
+  async openResourceFieldsModal() {
     if (this.loading) {
       return;
     }
     this.loading = true;
-    this.changeDetectorRef.detectChanges();
+    const { ConfigDisplayGridFieldsModalComponent } = await import(
+      './config-display-grid-fields-modal/config-display-grid-fields-modal.component'
+    );
     this.apollo
       .query<ResourceQueryResponse>({
         query: GET_SHORT_RESOURCE_BY_ID,
@@ -85,37 +90,35 @@ export class ResourceAvailableFieldsComponent
         },
       })
       .subscribe({
-        next: async ({ data }) => {
+        next: ({ data }) => {
           this.loading = false;
           if (data.resource && data.resource.name) {
             const nameTrimmed = data.resource.name
               .replace(/\s/g, '')
               .toLowerCase();
-            const { ConfigDisplayGridFieldsModalComponent } = await import(
-              '../../../components/config-display-grid-fields-modal/config-display-grid-fields-modal.component'
-            );
-            const dialogRef = this.dialog.open(
-              ConfigDisplayGridFieldsModalComponent,
-              {
-                data: {
-                  form: !this.model.obj.gridFieldsSettings
-                    ? null
-                    : this.convertFromRawToFormGroup(
-                        this.model.obj.gridFieldsSettings
-                      ),
-                  resourceName: nameTrimmed,
-                },
-              }
-            );
-            dialogRef.componentInstance?.ngOnInit();
-            this.changeDetectorRef.detectChanges();
-            dialogRef.closed
-              .pipe(takeUntil(this.destroy$))
-              .subscribe((res: any) => {
-                if (res && res.value.fields) {
-                  this.model.obj.gridFieldsSettings = res.getRawValue();
+            // Run dialog opening inside the current angular zone in order to keep the change detection correctly synchronized
+            this.ngZone.run(() => {
+              const dialogRef = this.dialog.open(
+                ConfigDisplayGridFieldsModalComponent,
+                {
+                  data: {
+                    form: !this.model.obj.gridFieldsSettings
+                      ? null
+                      : this.convertFromRawToFormGroup(
+                          this.model.obj.gridFieldsSettings
+                        ),
+                    resourceName: nameTrimmed,
+                  },
                 }
-              });
+              );
+              dialogRef.closed
+                .pipe(takeUntil(this.destroy$))
+                .subscribe((res: any) => {
+                  if (res && res.value.fields) {
+                    this.model.obj.gridFieldsSettings = res.getRawValue();
+                  }
+                });
+            });
           }
         },
         error: () => (this.loading = false),
